Build quota plugin on stable-3.10 and master vs. stable-3.13
The quota master branch is compatible with Gerrit v3.11,
v3.12 and v3.13. For building with Gerrit v3.10, a specific
stable-3.10 needs to be used for preventing breakages.

This amendment is needed for allowing a green build of
the quota plugin on stable-3.10 before the final archiving
on the EOL release.

Change-Id: I846cc0d9ed3310b900f1905573e0700743c9f6e8