Set the absolute build timeout to 1h for Gerrit validations
When the build was looping continuously, the job was kept alive
indefinitely because the previous timeout strategy was based
on the timeout on the output stream.

The previous timeout strategy caused the CI builds to
starvation due to jobs keeping the workers busy for hours.

Change the build timout strategy to absolute and set 60 mins
as the maximum time that a worker can be busy building a change,
so that any possible issue with tests or other build tasks
won't block all the workers indefinitely.

Change-Id: Ia22e27c0f436b5d4e232e45ede4e4fced0935f3c