Pin Docker client to v24.0.9 for compatibility Api 1.43
Existing CI/CD builds do rely on the Docker client to the server
communication that is based on Docker Api 1.43.

Keep the v24.0.9 client so that existing builds can continue to succeed
with the existing Docker server.

Also, avoid installing the engine because the CI/CD builds would just
need the client to connect to the remote Docker engine.

Change-Id: Id1bdea245bfc9b65b63336dd8759aa5a49bd01cb